rename createGeneralSlise to proper name
